refactor : 패키지 구조 리팩토링(#WA-43)
초반에 설계시 작성했던 web, application, persistence로 이루어진 패키지 구조를 리팩토링합니다.

Related to : WA-43
